requirements_dev.txt
MacFSEvents==0.4
colorama==0.3.2
cov-core==1.14.0
coverage==3.7.1
nose==1.3.4
py==1.4.26
pytest==2.6.4
python-dateutil==2.2
python-termstyle==0.1.10
redis==2.10.3
six==1.8.0
sniffer==0.3.3
test-pkg==0.0
wsgiref==0.1.2
